+++ frontend/src/routes/map/utils/map/style-json.ts
@@ -1,0 +1,8 @@
+/* style.jsonを取得 */
+export const getFieldDictJson = async (url: string): Promise<{ [key: string]: string }> => {
+	const response = await fetch(url);
+	if (!response.ok) {
+		throw new Error(`Failed to fetch ${url}`);
+	}
+	return await response.json();
+};
